Dev | Изменение UserSettings

+++ AutoCareDiray/ViewModels/MainPage/UserSettingsViewModal.cs
@@ -1,6 +1,7 @@
 ﻿using AutoCareDiray.Models;
 using AutoCareDiray.Service;
 using AutoCareDiray.View;
+using CommunityToolkit.Maui.Alerts;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using System;
@@ -14,6 +15,7 @@ namespace AutoCareDiray.ViewModels
     public partial class UserSettingsViewModal : ObservableObject
     {
         private readonly IApiService _apiService;
+        private readonly IDialogService _dialogService;
 
         [ObservableProperty]
         public string nickName;
@@ -22,12 +24,13 @@ namespace AutoCareDiray.ViewModels
         [ObservableProperty]
         public string user_id;
 
-        public UserSettingsViewModal(IApiService apiService) 
+        public UserSettingsViewModal(IApiService apiService, IDialogService dialogService) 
         {
             _apiService = apiService;
+            _dialogService = dialogService;
             NickName = Preferences.Get("NickName","null");
             Email = Preferences.Get("Email", "null");
-            User_id = Preferences.Get("User_id", "00");
+            User_id = Preferences.Get("User_id", "0");
         }
 
 
@@ -39,31 +42,29 @@ namespace AutoCareDiray.ViewModels
             Preferences.Remove("Email");
             Preferences.Remove("is_login");
 
-            var authPage = Application.Current.Handler.MauiContext.Services.GetService<AuthorizationPage>();
-            await Shell.Current.Navigation.PushModalAsync(authPage);
+            await Shell.Current.GoToAsync("//AuthorizationPage");
 
         }
         [RelayCommand]
         public async void DeleteProfil()
         {
-            var UserIdString = Preferences.Get("User_id", null);
-            if (int.TryParse(UserIdString, out var UserId))
+            var result = await _dialogService.ShowConfirmationMessage("Вы точно хотите удалить пользователя?");
+            if (result == true)
             {
-                var response = await _apiService.DeleteUserApiAsync(UserId);
-                if (response.Success == true)
+                var UserIdString = Preferences.Get("User_id", null);
+                if (int.TryParse(UserIdString, out var UserId))
                 {
-                    Preferences.Remove("User_Id");
-                    Preferences.Remove("NickName");
-                    Preferences.Remove("Email");
-                    Preferences.Remove("is_login");
+                    var response = await _apiService.DeleteUserApiAsync(UserId);
+                    if (response.Success == true)
+                    {
+                        Preferences.Remove("User_Id");
+                        Preferences.Remove("NickName");
+                        Preferences.Remove("Email");
+                        Preferences.Remove("is_login");
+                    }
+
+                    await Shell.Current.GoToAsync("//AuthorizationPage");
                 }
-
-                var authPage = Application.Current.Handler.MauiContext.Services.GetService<AuthorizationPage>();
-                await Shell.Current.Navigation.PushModalAsync(authPage);
-            }
-            else
-            {
-
             }
         }
     }
